Plot widgets interface with Qwt and Qwt3D (still ongoing). Plot data processing (still VTK dependent). Application paths.
*** Application paths in settings dialog. "Off gui" error dialog ***

ApplicationSettingsDlg.cpp:
+ValidateAndAssign: Now the user only has to deal with one message box instead of two or more. Same behavior as before but now the user is just warned about missing directories and a message box appears and asks if the user wants to create them.
+Constructor: Added code to access the application paths object inside settings object and loads (and displays) the current user paths.
+Fixed checkbox behavior on selection/de-selection to gray-out user base path if needed.

QtUtilsIO.h
+Added function SafeDuplicateFile (NOT used elsewhere, just added because It can be useful in situations such as RasterLayer copy procedure which is extremely slow and nothing happens with the previously chosen call to DuplicateFile because the destination path directory is inexistent.

main.cpp:
+Replaced the code related to error message box with a call to a static method from CBratApplication::OffGuiErrorMessage()

BratApplication.h:
+Added OffGuiErrorMessage signature

BratApplication.cpp:
+Added OffGuiErrorMessage implementation

ApplicationSettingsDlg.cpp:
+Added code to validate paths and / or create them. The paths inside ApplicationPaths from ApplicationSettings are updated accordingly.
+Unsure, but decided to throw an exception if we try to create directories (in case they don't exist) , this exception must eventually be caught outside (again, I was not so sure about this)

ApplicationPaths.h:
+Changed function name from SetUniqueUserBasePath to just SetUserBasePath (unique or not, it shall be decided by the first argument boolean)

ApplicationPaths.c++:
+Fixed the validation code on the ApplicationPaths constructor (!IsValid() instead of IsValid()) this is because we choose not to continue if either superclass validation failed (isValid() is false) or our subclass validationfailed (ValidatePaths() is false).
+Fixed the true/false thing on path validation under CApplicationPaths::ValidPaths()
+Recoded part of "SetUserBasePath( bool unique, const std::string &path ) (now this function should be able to set a new userbase path unique or non unique).

*** Plot widgets interface with Qwt and Qwt3D (started). Plot data processing (still VTK dependent). ***

- 2D and 3D plot raw data view code import from v3 into brat.
- work on 2D and 3D plotting data, still vtk dependent
- Qwt interface
- Qwt3d interface
- consolidation of 2D and 3D plot widgets (continued).
- mac and linux adjustments, as well as adjustments to old applications and in general to all projects, VS and qtc, for all platforms.

*** Coding standards updated. ***

+++ display/PlotData/PlotField.cpp
@@ -80,7 +80,7 @@ CPlotField* CPlotField::GetPlotField(CBratObject* ob)
 //----------------------------------------
 CInternalFiles* CPlotField::GetInternalFiles(int32_t index)
 {
-  int32_t nCount = m_internalFiles.size();
+  int32_t nCount = (int32_t)m_internalFiles.size();
 
   if ( (index < 0) || (index >= nCount))
   {
@@ -111,7 +111,7 @@ CInternalFiles* CPlotField::GetInternalFiles(int32_t index)
 //----------------------------------------
 CInternalFilesYFX* CPlotField::GetInternalFilesYFX(int32_t index)
 {
-  int32_t nCount = m_internalFiles.size();
+  int32_t nCount = (int32_t)m_internalFiles.size();
 
   CInternalFilesYFX* yfx = dynamic_cast<CInternalFilesYFX*>(GetInternalFiles(index));
   if (yfx == NULL)
